Reference

AIHelpMe.aihelp

AIHelpMe.docdata_to_source

AIHelpMe.docextract

AIHelpMe.docextract

AIHelpMe.docextract

AIHelpMe.docextract

AIHelpMe.find_new_chunks

AIHelpMe.last_context

AIHelpMe.load_index!

AIHelpMe.load_index!

AIHelpMe.preview_context

AIHelpMe.update_index

PromptingTools.Experimental.RAGTools.build_index

PromptingTools.Experimental.RAGTools.build_index

AIHelpMe.@aihelp!_str

AIHelpMe.@aihelp_str

AIHelpMe.aihelp — Method

aihelp([index::RAG.AbstractChunkIndex,]
    question::AbstractString;
    rag_template::Symbol = :RAGAnswerFromContext,
    top_k::Int = 100, top_n::Int = 5,
    minimum_similarity::AbstractFloat = -1.0,
    maximum_cross_similarity::AbstractFloat = 1.0,
    rerank_strategy::RAG.RerankingStrategy = (!isempty(PT.COHERE_API_KEY) ?
                                              RAG.CohereRerank() : RAG.Passthrough()),
    annotate_sources::Bool = true,
    model_embedding::String = PT.MODEL_EMBEDDING, model_chat::String = PT.MODEL_CHAT,
    chunks_window_margin::Tuple{Int, Int} = (1, 1),
    return_context::Bool = false, verbose::Integer = 1,
    rerank_kwargs::NamedTuple = NamedTuple(),
    api_kwargs::NamedTuple = NamedTuple(),
    kwargs...)

Generates a response for a given question using a Retrieval-Augmented Generation (RAG) approach over Julia documentation.

Arguments

index::AbstractChunkIndex: The chunk index (contains chunked and embedded documentation).

question::AbstractString: The question to be answered.

rag_template::Symbol: Template for the RAG model, defaults to :RAGAnswerFromContext.

top_k::Int: Number of top candidates to retrieve based on embedding similarity.

top_n::Int: Number of candidates to return after reranking. This is how many will be sent to the LLM model.

minimum_similarity::AbstractFloat: Minimum similarity threshold (between -1 and 1) for filtering chunks based on embedding similarity. Defaults to -1.0.

maximum_cross_similarity::AbstractFloat: Maximum cross-similarity threshold to avoid sending duplicate documents. NOT IMPLEMENTED YET

rerank_strategy::RerankingStrategy: Strategy for reranking the retrieved chunks. Defaults to Passthrough() or CohereRerank depending on whether COHERE_API_KEY is set.

model_embedding::String: Model used for embedding the question, default is PT.MODEL_EMBEDDING.

model_chat::String: Model used for generating the final response, default is PT.MODEL_CHAT.

chunks_window_margin::Tuple{Int,Int}: The window size around each chunk to consider for context building. See ?build_context for more information.

return_context::Bool: If true, returns the context used for RAG along with the response.

return_all::Bool: If true, returns all messages in the conversation (helpful to continue conversation later).

verbose::Bool: If true, enables verbose logging.

rerank_kwargs: Reranking parameters that will be forwarded to the reranking strategy

api_kwargs: API parameters that will be forwarded to the API calls

Returns

If return_context is false, returns the generated message (msg).

If return_context is true, returns a tuple of the generated message (msg) and the RAG context (rag_context).

Notes

The function first finds the closest chunks of documentation to the question (via embeddings).

It reranks the candidates and builds a "context" for the RAG model (ie, information to be provided to the LLM model together with the user question).

The chunks_window_margin allows including surrounding chunks for richer context, considering they are from the same source.

The function currently supports only single ChunkIndex.

Function always saves the last context in global LAST_CONTEXT for inspection of sources/context regardless of return_context value.

Examples

Using aihelp to get a response for a question:

index = build_index(...)  # create an index that contains Makie.jl documentation
question = "How to make a barplot in Makie.jl?"
msg = aihelp(index, question)

# or simply
msg = aihelp(index; question)

AIHelpMe.docdata_to_source — Method

docdata_to_source(data::AbstractDict)

Creates a source path from a given DocStr record

AIHelpMe.docextract — Function

docextract(d::DocStr, sep::AbstractString = "

")

Extracts the documentation from a DocStr record. Separates the individual docs within DocStr with sep.

AIHelpMe.docextract — Function

docextract(d::MultiDoc, sep::AbstractString = "

")

Extracts the documentation from a MultiDoc record (separates the individual docs within DocStr with sep)

AIHelpMe.docextract — Function

docextract(modules::Vector{Module} = Base.Docs.modules)

Extracts the documentation from a vector of modules.

AIHelpMe.docextract — Method

docextract(mod::Module)

Extracts the documentation from a given (loaded) module.

AIHelpMe.find_new_chunks — Method

find_new_chunks(old_chunks::AbstractVector{<:AbstractString},
    new_chunks::AbstractVector{<:AbstractString})

Identifies the new chunks in new_chunks that are not present in old_chunks.

Returns a mask of chunks that are new (not present in old_chunks).

Uses SHA256 hashes to dedupe the strings quickly and effectively.

AIHelpMe.last_context — Method

last_context()

Returns the RAGContext from the last aihelp call. It can be useful to see the sources/references used by the AI model to generate the response.

If you're using aihelp() make sure to set return_context = true to return the context.

AIHelpMe.load_index! — Function

load_index!(file_path::Union{Nothing, AbstractString} = nothing;
    verbose::Bool = true, kwargs...)

Loads the serialized index in file_path into the global variable MAIN_INDEX. If not provided, it will download the latest index from the AIHelpMe.jl repository (more cost-efficient).

AIHelpMe.load_index! — Method

load_index!(index::RAG.AbstractChunkIndex;
    verbose::Bool = 1, kwargs...)

Loads the provided index into the global variable MAIN_INDEX.

If you don't have an index yet, use build_index to build one from your currently loaded packages (see ?build_index)

Example

# build an index from some modules, keep empty to embed all loaded modules (eg, `build_index()`) 
index = AIH.build_index([DataFramesMeta, DataFrames, CSV])
AIH.load_index!(index)

AIHelpMe.preview_context — Function

preview_context(context = last_context())

Preview the context of the last aihelp call. It will pretty-print the question, context and answer in the REPL.

AIHelpMe.update_index — Function

update_index(index::RAG.AbstractChunkIndex = MAIN_INDEX,
    modules::Vector{Module} = Base.Docs.modules;
    verbose::Integer = 1,
    separators = ["\n\n", ". ", "\n"], max_length::Int = 256,
    model::AbstractString = PT.MODEL_EMBEDDING,
    kwargs...)
    modules::Vector{Module} = Base.Docs.modules;
    verbose::Bool = true, kwargs...)

Updates the provided index with the documentation of the provided modules.

Deduplicates against the index.sources and embeds only the new document chunks (as measured by a hash).

Returns the updated index (new instance).

Example

If you loaded some new packages and want to add them to your MAIN_INDEX (or any index you use), run:

# To update the MAIN_INDEX
AHM.update_index() |> AHM.load_index!

# To update an explicit index
index = AHM.update_index(index)

PromptingTools.Experimental.RAGTools.build_index — Function

RAG.build_index(modules::Vector{Module} = Base.Docs.modules; verbose::Int = 1,
    separators = ["

", ". ", " "], max_length::Int = 256, kwargs...)

Build index from the documentation of the currently loaded modules. If modules is empty, it will use all currently loaded modules.

PromptingTools.Experimental.RAGTools.build_index — Method

RAG.build_index(mod::Module; verbose::Int = 1, kwargs...)

Build index from the documentation of a given module mod.

AIHelpMe.@aihelp!_str — Macro

aihelp!"user_question"[model_alias] -> AIMessage

The aihelp!"" string macro is used to continue a previous conversation with the AI model.

It appends the new user prompt to the last conversation in the tracked history (in AIHelpMe.CONV_HISTORY) and generates a response based on the entire conversation context. If you want to see the previous conversation, you can access it via AIHelpMe.CONV_HISTORY, which keeps at most last PromptingTools.MAX_HISTORY_LENGTH conversations.

It does NOT provide new context from the documentation. To do that, start a new conversation with aihelp"<question>".

Arguments

user_question (String): The follow up question to be added to the existing conversation.

model_alias (optional, any): Specify the model alias of the AI model to be used (see PT.MODEL_ALIASES). If not provided, the default model is used.

Returns

AIMessage corresponding to the new user prompt, considering the entire conversation history.

Example

To continue a conversation:

# start conversation as normal
aihelp"How to create a dictionary?" 

# ... wait for reply and then react to it:

# continue the conversation (notice that you can change the model, eg, to more powerful one for better answer)
aihelp!"Can you create it from named tuple?"gpt4t
# AIMessage("Yes, you can create a dictionary from a named tuple ...")

Usage Notes

This macro should be used when you want to maintain the context of an ongoing conversation (ie, the last ai"" message).

It automatically accesses and updates the global conversation history.

If no conversation history is found, it raises an assertion error, suggesting to initiate a new conversation using ai"" instead.

Important

Ensure that the conversation history is not too long to maintain relevancy and coherence in the AI's responses. The history length is managed by MAX_HISTORY_LENGTH.

AIHelpMe.@aihelp_str — Macro

aihelp"user_question"[model_alias] -> AIMessage

The aihelp"" string macro generates an AI response to a given user question by using aihelp under the hood. It will automatically try to provide the most relevant bits of the documentation (from the index) to the LLM to answer the question.

See also aihelp!"" if you want to reply to the provided message / continue the conversation.

Arguments

user_question (String): The question to be answered by the AI model.

model_alias (optional, any): Provide model alias of the AI model (see MODEL_ALIASES).

Returns

AIMessage corresponding to the input prompt.

Example

result = aihelp"Hello, how are you?"
# AIMessage("Hello! I'm an AI assistant, so I don't have feelings, but I'm here to help you. How can I assist you today?")

If you want to interpolate some variables or additional context, simply use string interpolation:

a=1
result = aihelp"What is `$a+$a`?"
# AIMessage("The sum of `1+1` is `2`.")

If you want to use a different model, eg, GPT-4, you can provide its alias as a flag:

result = aihelp"What is `1.23 * 100 + 1`?"gpt4t
# AIMessage("The answer is 124.")